Remove under performing constructor argument analyzer for interfaces
